requirements.txt
Django==5.2.8
psycopg2-binary==2.9.11
python-dotenv==1.2.1
dj-database-url==3.0.1
whitenoise==6.11.0
gunicorn==23.0.0
django-crispy-forms==2.1
crispy-bootstrap4==2023.1
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
google-cloud-firestore==2.17.0